# Conflicts:
#	README.md

+++ src/main/java/com/github/kongchen/swagger/docgen/AbstractDocumentSource.java
@@ -21,6 +21,7 @@ import io.swagger.converter.ModelConverters;
 import io.swagger.models.Scheme;
 import io.swagger.models.Swagger;
 import io.swagger.models.properties.Property;
+import io.swagger.util.Json;
 import io.swagger.util.Yaml;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
@@ -95,8 +96,12 @@ public abstract class AbstractDocumentSource {
 
 
     public abstract void loadDocuments() throws GenerateException;
-
+    
     public void toSwaggerDocuments(String uiDocBasePath, String outputFormats) throws GenerateException {
+        toSwaggerDocuments(uiDocBasePath, outputFormats, null);
+    }
+
+    public void toSwaggerDocuments(String uiDocBasePath, String outputFormats, String fileName) throws GenerateException {
         mapper.configure(SerializationFeature.WRITE_EMPTY_JSON_ARRAYS, false);
         mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
 
@@ -123,8 +128,10 @@ public abstract class AbstractDocumentSource {
                 throw new GenerateException(String.format("Create Swagger-outputDirectory[%s] failed.", swaggerPath));
             }
         }
-        cleanupOlds(dir);
-
+        cleanupOlds(dir, outputFormats);
+        if (fileName == null || "".equals(fileName.trim())) {
+            fileName = "swagger";
+        }
         try {
             if (outputFormats != null) {
                 for (String format : outputFormats.split(",")) {
@@ -133,10 +140,10 @@ public abstract class AbstractDocumentSource {
                         switch (output) {
                             case json:
                                 ObjectWriter jsonWriter = mapper.writer(new DefaultPrettyPrinter());
-                                FileUtils.write(new File(dir, "swagger.json"), jsonWriter.writeValueAsString(swagger));
+                                FileUtils.write(new File(dir, fileName + ".json"), jsonWriter.writeValueAsString(swagger));
                                 break;
                             case yaml:
-                                FileUtils.write(new File(dir, "swagger.yaml"), Yaml.pretty().writeValueAsString(swagger));
+                                FileUtils.write(new File(dir, fileName + ".yaml"), Yaml.pretty().writeValueAsString(swagger));
                                 break;
                         }
                     } catch (Exception e) {
@@ -146,7 +153,7 @@ public abstract class AbstractDocumentSource {
             } else {
                 // Default to json
                 ObjectWriter jsonWriter = mapper.writer(new DefaultPrettyPrinter());
-                FileUtils.write(new File(dir, "swagger.json"), jsonWriter.writeValueAsString(swagger));
+                FileUtils.write(new File(dir, fileName + ".json"), jsonWriter.writeValueAsString(swagger));
             }
         } catch (IOException e) {
             throw new GenerateException(e);
@@ -154,7 +161,7 @@ public abstract class AbstractDocumentSource {
     }
 
     public void loadModelModifier() throws GenerateException, IOException {
-        ObjectMapper objectMapper = new ObjectMapper();
+        ObjectMapper objectMapper = Json.mapper();
         if (apiSource.isUseJAXBAnnotationProcessor()) {
             objectMapper.registerModule(new JaxbAnnotationModule());
             objectMapper.registerModule(new JaxbAnnotationModule());
@@ -234,11 +241,13 @@ public abstract class AbstractDocumentSource {
     }
 
 
-    private void cleanupOlds(File dir) {
-        if (dir.listFiles() != null) {
-            for (File f : dir.listFiles()) {
-                if (f.getName().endsWith("json")) {
-                    f.delete();
+    private void cleanupOlds(File dir, String outputFormats) {
+        if (dir.listFiles() != null && outputFormats != null) {
+            for (String format : outputFormats.split(",")) {
+                for (File f : dir.listFiles()) {
+                    if (f.getName().endsWith(format.toLowerCase())) {
+                        f.delete();
+                    }
                 }
             }
         }
